+++ src/main/handlers/__init__.py
@@ -13,7 +13,7 @@ from tasks.lesson03 import task307
 from tasks.lesson03 import task309
 from tasks.lesson03 import task310
 from tasks.lesson03 import task311
-from tasks.lesson04 import task402
+
 
 from . import error_test
 from . import index
@@ -22,7 +22,7 @@ from .system_handlers import handle_500
 
 urlpatterns: Dict[str, HandlerT] = {
     "/": index.handler,
-    "/api/v1/tasks/402/": task402.handler_api,
+    "/api/v1/tasks/402/": task4022.handler_api,
     "/e/": error_test.handler,
     "/tasks/1/103/": task103.handler,
     "/tasks/3/301/": task301.handler,
@@ -35,7 +35,7 @@ urlpatterns: Dict[str, HandlerT] = {
     "/tasks/3/309/": task309.handler,
     "/tasks/3/310/": task310.handler,
     "/tasks/3/311/": task311.handler,
-    "/tasks/4/402/": task402.handler,
+    "/tasks/4/402/": task4022.handler,
 }
