+++ src/main/templates/index.html
@@ -36,7 +36,7 @@
             <li>
                 <p>Урок 4</p>
                 <ul>
-                    <li><a href="/tasks/402/">Задача 4.02</a></li>
+                    <li><a href="/tasks/lesson04/task402.html/">Задача 4.02</a></li>
                 </ul>
             </li>
         </ul>
