+++ src/main/templates/tasks/lesson04/task402.html
@@ -110,3 +110,4 @@
 </script>
 </body>
 </html>
+
