+++ src/tasks/lesson03/task306.py
@@ -27,7 +27,7 @@ def handler(request: HttpRequest) -> HttpResponse:
 
     document = render_template(TEMPLATE, context)
 
-    response = HttpResponse(payload=document)
+    response = HttpResponse(document)
 
     return response
