+++ src/tasks/lesson03/task310.py
@@ -46,7 +46,7 @@ def handler(request: HttpRequest) -> HttpResponse:
 
     document = render_template(TEMPLATE, context)
 
-    response = HttpResponse(payload=document)
+    response = HttpResponse(document)
 
     return response
