+++ src/tasks/lesson03/task311.py
@@ -25,7 +25,7 @@ def handler(request: HttpRequest) -> HttpResponse:
 
     document = render_template(TEMPLATE, context)
 
-    response = HttpResponse(payload=document)
+    response = HttpResponse(document)
 
     return response
